added cart apis

+++ model/Cart.js
@@ -5,12 +5,12 @@ let Cart = sequelizeInstance.define(
   "cart",
   {
     id: {
-      type: sequelize.DataTypes.BIGINT,
+      type: sequelize.DataTypes.INTEGER,
       primaryKey: true,
       autoIncrement: true,
     },
     cost: {
-      type: sequelize.DataTypes.INTEGER,
+      type: sequelize.DataTypes.DECIMAL,
     },
   },
   {
